Refactor: UserController 네이밍 변경
